feat: Add stats card component
