+++ apps/admin-panel/src/components/KPI/index.ts
@@ -1,0 +1,1 @@
+export * from "./StatsCard";
